PYTTB_UTILS: Pull out utility only used internally in sptensor
