+++ tests/test_pyttb_utils.py
@@ -52,42 +52,6 @@ def test_sptensor_from_dense_matrix():
 
 
 @pytest.mark.indevelopment
-def test_sptensor_to_sparse_matrix():
-    subs = np.array([[1, 1, 1], [1, 1, 3], [2, 2, 2], [3, 3, 3]])
-    vals = np.array([[0.5], [1.5], [2.5], [3.5]])
-    shape = (4, 4, 4)
-    mode0 = sparse.coo_matrix(([0.5, 1.5, 2.5, 3.5], ([5, 13, 10, 15], [1, 1, 2, 3])))
-    mode1 = sparse.coo_matrix(([0.5, 1.5, 2.5, 3.5], ([5, 13, 10, 15], [1, 1, 2, 3])))
-    mode2 = sparse.coo_matrix(([0.5, 1.5, 2.5, 3.5], ([5, 5, 10, 15], [1, 3, 2, 3])))
-    Ynt = [mode0, mode1, mode2]
-    sptensorInstance = ttb.sptensor().from_data(subs, vals, shape)
-
-    for mode in range(sptensorInstance.ndims):
-        Xnt = ttb.tt_to_sparse_matrix(sptensorInstance, mode, True)
-        assert (Xnt != Ynt[mode]).nnz == 0
-        assert Xnt.shape == Ynt[mode].shape
-
-
-@pytest.mark.indevelopment
-def test_sptensor_from_sparse_matrix():
-    subs = np.array([[1, 1, 1], [1, 1, 3], [2, 2, 2], [3, 3, 3]])
-    vals = np.array([[0.5], [1.5], [2.5], [3.5]])
-    shape = (4, 4, 4)
-    sptensorInstance = ttb.sptensor().from_data(subs, vals, shape)
-    for mode in range(sptensorInstance.ndims):
-        sptensorCopy = ttb.sptensor.from_tensor_type(sptensorInstance)
-        Xnt = ttb.tt_to_sparse_matrix(sptensorCopy, mode, True)
-        Ynt = ttb.tt_from_sparse_matrix(Xnt, sptensorCopy.shape, mode, 0)
-        assert sptensorCopy.isequal(Ynt)
-
-    for mode in range(sptensorInstance.ndims):
-        sptensorCopy = ttb.sptensor.from_tensor_type(sptensorInstance)
-        Xnt = ttb.tt_to_sparse_matrix(sptensorCopy, mode, False)
-        Ynt = ttb.tt_from_sparse_matrix(Xnt, sptensorCopy.shape, mode, 1)
-        assert sptensorCopy.isequal(Ynt)
-
-
-@pytest.mark.indevelopment
 def test_tt_union_rows():
     a = np.array([[4, 6], [1, 9], [2, 6], [2, 6], [99, 0]])
     b = np.array([[1, 7], [1, 8], [2, 6]])
